# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "myAgenci.ai Mobile App API",
        "description": "REST API for myAgenci.ai Mobile Application.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.47:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/mobile/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard data",
                "description": "Returns KPIs, pipeline funnel, source distribution, financials, today's follow-ups, pending reminders, recent leads, branch performance, team performance, and 6-month trend. All sections respect the same filters as the web dashboard.",
                "operationId": "86cc9c0b8350f0b97b74075e266f631c",
                "parameters": [
                    {
                        "name": "branch_id",
                        "in": "query",
                        "description": "Filter by branch ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "stage",
                        "in": "query",
                        "description": "Filter by lead stage",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "new",
                                "qualified",
                                "proposal",
                                "negotiation",
                                "won",
                                "lost"
                            ]
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter by lead source",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "reference",
                                "ad_campaign",
                                "direct_visit",
                                "invitation",
                                "cold_outreach",
                                "social_media",
                                "website"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Start date (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "End date (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "quick_date",
                        "in": "query",
                        "description": "Preset date range",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "today",
                                "week",
                                "month",
                                "quarter",
                                "year"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard data fetched."
                                        },
                                        "data": {
                                            "description": "See schema below",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}